Don't use fzf for commandline autocomplete
It's too fuzzy and doesn't handle exact prefix match in many cases.

For example: if we have `autopilot.py` and `test_autopilot.py` in the
same directory, and we type `au<Tab>`, with the default complete, it
will match `autopilot.py`, but with fzf, it shows 2 matches and we need
an additional ENTER to select the first one.