Replace git-messenger with FzfLCommits
git-messenger has a bug where it doesn't update the line number when
running the blame command when we go back in time (using `o`). I
couldn't find a commit that introduced a change because of that.

Also, since we removed git-messenger, `gm` now will be used to submit
MR, since it's easier to reach than `g\`.