Add HapticKey
Since Touch Bar is just a flat panel, there is no feedback other than a
visual change of the key displayed, it is terribly uncomfortable when
tapping keys there. By adding a haptic feedback (which is vibrating a
trackpad not Touch Bar itself) it helps to improve the usage of Touch
Bar.

https://github.com/niw/HapticKey