A more responsive ENTER key
Previously, I map ENTER key to CONTROL when holding, but it caused a
delay of 250ms to the ENTER key. Removing this mapping forces me to use
the CONTROL key on the left, but it would speed up command run & file
opening, which is more important than having an addition CONTROL key on
the right.