refactor(tmux): make tmux and kitty play nice together
This allows using mouse to scroll tmux's scrollback without enabling
tmux mouse mode. This provides a better UX on select to copy & text
selection, since it will be handled natively by kitty.

I don't like tmux mouse mode, because the cursor will be changed to a
pointer, and making it works similar to kitty is impossible. So, let's
just allow kitty to handle the mouse actions.