feat(kitty): `ctrl+t` to open a new split with cwd
Because I run everything in a vagrant box, this feature didn't work
before, because Kitty doesn't know about the cwd on a remote server.

But Kitty v0.25.0 has introduced `Truly convenient SSH`, which makes
`new_window_with_cwd` works with remote hosts when shell integration is
on.

I have set it up and it does work as advertised.