bump(remote): upgrade fish to v3.6.1
It was not convenient to build fish x86_64 now, since I'm using a M1
MacBook (we can do that using buildx Docker Desktop, but I don't want to
leave my VM). The fish built that way also didn't work when running on
remote servers (`error while loading shared libraries: libpcre2-32.so`).

This switches to use the appimage version. It is bigger, but will be
easier to update in the future.