chore(tmux): switch default cursor shape back to arrow
It allows to select text more precisely, since tmux selection doesn't
work exactly as kitty selection. The first char when we start the
selection, tmux selects it if we point the cursor slightly over it,
but kitty only selects if we point cursor more than half of the char
width.