fix(tmux): unable to use mouse to select tab
This enable tmux mouse mode by default all the time, but with a slight
different compared to the previous implementation: we'll exit the copy
mode only if the selected text is in the last screen (we haven't
scrolled up).

The Kitty's text selection behavior will be overridden by tmux's
behavior. The off-by-one issue is still there but it's less annoying
than not being able to select tabs with mouse.