feat!(macOS): remove alt-tab
Since macOS Sonoma, it keeps showing an icon saying `AltTab is capturing
your screen`. While this is expected because AltTab needs to do that to
show thumbnails, but the icon distractes me and I don't want to lose
focus on the things I'm doing because of that. I can live without
AltTab, so let's remove it.